Crate libcrux_hmac

HMAC

This crate implements HMAC on SHA 1 and SHA 2 (except for SHA 224).

Enums

Algorithm

The HMAC algorithm defining the used hash function.

Functions

hmac

Compute the HMAC value with the given alg and key on data with an output tag length of tag_length. Returns a vector of length tag_length. Panics if either key or data are longer than u32::MAX.

hmac_sha1

Compute HMAC.

hmac_sha2_256

Compute HMAC.

hmac_sha2_384

Compute HMAC.

hmac_sha2_512

Compute HMAC.

tag_size

Get the tag size for a given algorithm.